pdf-stamp/backend/requirements.txt
flask
flask-cors
PyMuPDF
Pillow
openai
requests
